Revert 38748 - Merge 33672  Send resource load notifications from plugins to devtools agent. This is a part of a fix which would allow to inspect plugin resources.
BUG=2084

Review URL: http://codereview.chromium.org/435044

TBR=yurys@chromium.org
Review URL: http://codereview.chromium.org/597039

TBR=jam@chromium.org
Review URL: http://codereview.chromium.org/603027

git-svn-id: svn://svn.chromium.org/chrome/branches/249/src@38749 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/common/plugin_messages.h
@@ -54,7 +54,7 @@ struct PluginHostMsg_URLRequest_Params {
 };
 
 struct PluginMsg_URLRequestReply_Params {
-  unsigned long resource_id;
+  int resource_id;
   GURL url;
   bool notify_needed;
   intptr_t notify_data;
@@ -62,7 +62,7 @@ struct PluginMsg_URLRequestReply_Params {
 };
 
 struct PluginMsg_DidReceiveResponseParams {
-  unsigned long id;
+  int id;
   std::string mime_type;
   std::string headers;
   uint32 expected_length;
