+++ webkit/glue/webplugin_impl.h
@@ -128,16 +128,14 @@ class WebPluginImpl : public WebPlugin,
                              const char* url, Referrer referrer_flag);
 
   // Cancels a pending request.
-  void CancelResource(unsigned long id);
+  void CancelResource(int id);
 
-  // Returns the next avaiable resource id. Returns 0 if the operation fails.
-  // It may fail if the page has already been closed.
-  unsigned long GetNextResourceId();
+  // Returns the next avaiable resource id.
+  int GetNextResourceId();
 
   // Initiates HTTP GET/POST requests.
   // Returns true on success.
-  bool InitiateHTTPRequest(unsigned long resource_id,
-                           WebPluginResourceClient* client,
+  bool InitiateHTTPRequest(int resource_id, WebPluginResourceClient* client,
                            const char* method, const char* buf, int buf_len,
                            const GURL& url, const char* range_info,
                            Referrer referrer_flag);
@@ -205,7 +203,7 @@ class WebPluginImpl : public WebPlugin,
                                 intptr_t existing_stream, bool notify_needed,
                                 intptr_t notify_data);
 
-  void SetDeferResourceLoading(unsigned long resource_id, bool defer);
+  void SetDeferResourceLoading(int resource_id, bool defer);
 
   // Ignore in-process plugins mode for this flag.
   bool IsOffTheRecord() { return false; }
@@ -230,7 +228,7 @@ class WebPluginImpl : public WebPlugin,
   void OnDownloadPluginSrcUrl();
 
   struct ClientInfo {
-    unsigned long id;
+    int id;
     WebPluginResourceClient* client;
     WebKit::WebURLRequest request;
     bool pending_failure_notification;
